tests: Split demo and realm creation tests into separate file.
